Improve structure in protostuff-it: move proto files to separate folders according to used template
