+++ protostuff-it/src/test/java/io/protostuff/compiler/java_bean/RepeatedIT.java
@@ -1,9 +1,9 @@
-package io.protostuff.compiler;
+package io.protostuff.compiler.java_bean;
 
 import java.io.IOException;
 import java.util.Arrays;
 
-import io.protostuff.compiler.it.UnmodifiableInt32List;
+import io.protostuff.compiler.it.java_bean.UnmodifiableInt32List;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
@@ -12,7 +12,7 @@ import org.mockito.Mockito;
 
 import io.protostuff.Input;
 import io.protostuff.Schema;
-import io.protostuff.compiler.it.Int32List;
+import io.protostuff.compiler.it.java_bean.Int32List;
 
 /**
  * Integration tests for java_bean repeated fields
