+++ protostuff-it/src/test/proto/java_bean/EmptyMessageIT.proto
@@ -1,5 +1,5 @@
 package it;
-option java_package = "io.protostuff.compiler.it";
+option java_package = "io.protostuff.compiler.it.java_bean";
 
 message EmptyMessage {
   // Protostuff Compiler should generate valid code for messages without fields
