+++ protostuff-it/src/test/proto/java_bean_model/JavaBeanModelIT.proto
@@ -1,6 +1,6 @@
 package it;
 
-option java_package = "io.protostuff.compiler.it";
+option java_package = "io.protostuff.compiler.it.java_bean_model";
 option models = true;
 
 message JavaBeanModelSimpleMessage {
