+++ protostuff-it/src/test/proto/java_bean_primitives/PrimitivesIT.proto
@@ -1,5 +1,5 @@
 package it;
-option java_package = "io.protostuff.compiler.it";
+option java_package = "io.protostuff.compiler.it.java_bean_primitives";
 
 message PrimitiveFields {
   optional int32 a = 1;
